requirements.txt
fastapi
uvicorn
groq
sentence-transformers
faiss-cpu
nltk
python-dotenv
requests
streamlit
pandas
